% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUC.R
\name{AUC}
\alias{AUC}
\title{Area Under the Curve}
\usage{
AUC(measures)
}
\arguments{
\item{measures}{An object returned by the \code{roc} function.}
}
\value{
A vector containing the AUCs for each slot of the \code{roc} table.
}
\description{
Calculate the area under the ROC curve for the class \code{roc}.
}
