% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R, R/plot.R
\name{roc}
\alias{roc}
\alias{plot.roc}
\title{Binary Calibration}
\usage{
roc(data_list, model, target_name, cut_off_from = 1)

\method{plot}{roc}(x, interactive = TRUE, ...)
}
\arguments{
\item{data_list}{List of data tables each one suitable from the argument \code{newdata} of
the \code{predict.glm} function.}

\item{model}{A model of class \code{glm}.}

\item{target_name}{Character specifying the name of the column of each data table
containing the target variable.}

\item{cut_off_from}{Character or integer indicating the location into the \code{data_list}
from which candidate cut-offs must be extracted.}

\item{x}{As object resulting from the function \code{roc}.}

\item{interactive}{Make the plot interactive using \code{ggplotly}.}

\item{...}{Further arguments for the function \code{plot} (not implemented).}
}
\description{
Binary Calibration
}
\examples{
set.seed(666)
n = 50
tab <- data.frame(x=rnorm(n))
tab$y <- tab$x*2+rnorm(n,sd=2)
tab$y <- as.integer(exp(tab$y)/(1+exp(tab$y))>0.5)
tab$set <- gl(2, n/2, labels=c("train","test"))
tab <- split(tab, f=tab$set)
model <- glm(y~x, family=binomial, data=tab$train)
measures <- roc(tab, model=model, target_name="y")
plot(measures)
}
